requirements.txt
PyQt6==6.5.1
PyQt6-Qt6==6.5.2
PyQt6-sip==13.5.1
pyspellchecker==0.7.2
